added cpp interface to color button and split color items
