+++ DearPyGui/cmake/dpg_sources.cmake
@@ -83,7 +83,9 @@ set(MARVEL_SOURCES
 	"src/core/AppItems/basic/mvSliderItems.cpp"
 	"src/core/AppItems/basic/mvImage.cpp"
 	"src/core/AppItems/basic/mvImageButton.cpp"
-	"src/core/AppItems/basic/mvColorItems.cpp"
+	"src/core/AppItems/basic/mvColorButton.cpp"
+	"src/core/AppItems/basic/mvColorEdit.cpp"
+	"src/core/AppItems/basic/mvColorPicker.cpp"
 	"src/core/AppItems/basic/mvRadioButton.cpp"
 	"src/core/AppItems/basic/mvListbox.cpp"
 	"src/core/AppItems/basic/mvCombo.cpp"
