+++ DearPyGui/src/core/AppItems/mvAppItems.h
@@ -10,7 +10,9 @@
 #include "basic/mvInputItems.h"
 #include "basic/mvSliderItems.h"
 #include "basic/mvDragItems.h"
-#include "basic/mvColorItems.h"
+#include "basic/mvColorButton.h"
+#include "basic/mvColorEdit.h"
+#include "basic/mvColorPicker.h"
 #include "basic/mvText.h"
 #include "basic/mvTabButton.h"
 #include "basic/mvMenuItem.h"
