+++ DearPyGui/src/core/AppItems/nodes/mvNode.cpp
@@ -107,6 +107,9 @@ namespace Marvel {
 		m_ypos = pos.y;
 	}
 
+#ifdef MV_CPP
+#else
+
 	void mvNode::setExtraConfigDict(PyObject* dict)
 	{
 		if (dict == nullptr)
@@ -170,4 +173,5 @@ namespace Marvel {
 
 	}
 
+#endif
 }
