+++ DearPyGui/src/core/AppItems/nodes/mvNodeEditor.cpp
@@ -340,6 +340,9 @@ namespace Marvel {
 		
 	}
 
+#ifdef MV_CPP
+#else
+
 	PyObject* add_node_editor(PyObject* self, PyObject* args, PyObject* kwargs)
 	{
 		const char* name;
@@ -605,4 +608,6 @@ namespace Marvel {
 		return GetPyNone();
 
 	}
+
+#endif
 }
